testRegEx
At this point, this project is able to recognize some C++ code. Does not recognize comments or while loops.
Does something wierd when it reaches the end of the C++ code. It has trouble removing the last line of cppCode
and I am forced to use substring(1) to removed unrecognized characters one by one. This is not ideal and a fix will be needed.